astar_lookahead alone
