+++ astar_lookahead_publisher/astar_lookahead_pub.py
@@ -7,54 +7,25 @@ from geometry_msgs.msg import PoseStamped
 import math
 import csv
 import os
-from ackermann_msgs.msg import AckermannDriveStamped
+# from ackermann_msgs.msg import AckermannDriveStamped
 from visualization_msgs.msg import Marker, MarkerArray
 from geometry_msgs.msg import Point
 import numpy as np
-from std_msgs.msg import Bool
 from tf2_ros import Buffer, TransformListener
-from nav2_msgs.action import ComputePathToPose
-from rclpy.action import ActionClient
-
-def euler_from_quaternion(quaternion):
-    """
-    Converts quaternion (w in last place) to euler roll, pitch, yaw
-    quaternion = [x, y, z, w]
-    Bellow should be replaced when porting for ROS 2 Python tf_conversions is done.
-    """
-    x = quaternion[0]
-    y = quaternion[1]
-    z = quaternion[2]
-    w = quaternion[3]
-
-    sinr_cosp = 2 * (w * x + y * z)
-    cosr_cosp = 1 - 2 * (x * x + y * y)
-    roll = np.arctan2(sinr_cosp, cosr_cosp)
-
-    sinp = 2 * (w * y - z * x)
-    pitch = np.arcsin(sinp)
-
-    siny_cosp = 2 * (w * z + x * y)
-    cosy_cosp = 1 - 2 * (y * y + z * z)
-    yaw = np.arctan2(siny_cosp, cosy_cosp)
-
-    return roll, pitch, yaw
 
 class AstarLookahead(Node):
     def __init__(self):
         super().__init__("astar_lookahead_pub_node")
 
-        self.declare_parameter("is_antiClockwise", False)
         self.declare_parameter("lookahead_distance", 0.0)
         self.declare_parameter("lookahead_marker_topic", "")
-        self.declare_parameter("object_detected_topic", "/tmp/obj_detected")
         self.declare_parameter("csv_path", "")   
+        self.declare_parameter("astar_pp_path", "")
 
-        self.is_antiClockwise = self.get_parameter("is_antiClockwise").get_parameter_value().bool_value
         self.lookahead_distance = self.get_parameter("lookahead_distance").get_parameter_value().double_value
         self.marker_pub_topic = self.get_parameter("lookahead_marker_topic").get_parameter_value().string_value
-        self.obj_detected_topic = self.get_parameter("object_detected_topic").get_parameter_value().string_value
         self.csv_path = self.get_parameter("csv_path").get_parameter_value().string_value
+        self.astar_pp_path = self.get_parameter("astar_pp_path").get_parameter_value().string_value
 
         self.tf_buffer = Buffer()
         self.tf_listener = TransformListener(self.tf_buffer, self)
@@ -64,10 +35,7 @@ class AstarLookahead(Node):
 
         self.lookahead_marker_pub = self.create_publisher(Marker, self.marker_pub_topic, 10)
         self.lookahead_circle_pub = self.create_publisher(Marker, "/astar_lookahead_circle", 10)
-        self.path_publisher = self.create_publisher(Path, "/tmp/astar_pp_path", 10)
-        # self.obj_detected_sub = self.create_subscription(Bool, self.obj_detected_topic, self.create_path, 10)
-        self._action_client = ActionClient(self, ComputePathToPose, 'compute_path_to_pose')
-        self.goal_sent = False
+        self.path_publisher = self.create_publisher(Path, self.astar_pp_path, 10)
         self.marker = None
 
     def load_path_from_csv(self, csv_path):
@@ -78,35 +46,6 @@ class AstarLookahead(Node):
                 x, y, v = float(row[0]), float(row[1]), float(row[2])
                 path.append((x, y, v))
         return path
-
-    def create_path(self, msg: Bool): 
-        if not msg.data or not self.marker:
-            return
-        marker = self.marker
-        goal_msg = ComputePathToPose.Goal()
-
-        # Create a PoseStamped message for the goal
-        goal_pose = PoseStamped()
-        goal_pose.header.frame_id = marker.header.frame_id
-        goal_pose.header.stamp = self.get_clock().now().to_msg()
-        goal_pose.pose = marker.pose
-
-        # Assign the goal pose
-        goal_msg.goal = goal_pose
-
-        # Wait for action server and send goal
-        self._action_client.wait_for_server()
-        self._send_goal_future = self._action_client.send_goal_async(goal_msg)
-        self._send_goal_future.add_done_callback(self.goal_response_callback)
-        self.goal_sent = True
-
-    def path_update_cb(self, msg:Path):
-        self.path.clear() # to clear the path
-        for i in range(len(msg.poses)):
-            self.path.append((msg.poses[i].pose.position.x, msg.poses[i].pose.position.y, msg.poses[i].pose.orientation.w))
-        if self.is_antiClockwise:
-            self.path.reverse()
-        self.get_logger().info(f"path has been updated...")
 
     def get_pose(self):
         try:
@@ -119,13 +58,7 @@ class AstarLookahead(Node):
             )
 
             trans = transform.transform.translation
-            rot = transform.transform.rotation
-
-            # Convert quaternion to yaw
-            orientation_list = [rot.x, rot.y, rot.z, rot.w]
-            _, _, yaw = euler_from_quaternion(orientation_list)
-
-            # self.get_logger().info(f"Robot Pose - x: {trans.x:.2f}, y: {trans.y:.2f}, yaw: {yaw:.2f}")
+            
             x, y = trans.x, trans.y
             
             self.publish_lookahead_circle(x, y)
@@ -134,9 +67,6 @@ class AstarLookahead(Node):
                 self.get_logger().warn("No lookahead point found go ")
             else:
                 self.publish_lookahead_marker(lookahead_point)
-            msg = Bool()
-            msg.data = True
-            self.create_path(msg)
 
         except Exception as e:
             self.get_logger().warn(f"Transform not available: {e}")
@@ -200,32 +130,7 @@ class AstarLookahead(Node):
         marker.color.g = 0.0
         marker.color.b = 0.0
         self.lookahead_marker_pub.publish(marker)
-        if self.goal_sent:
-            return
-        
         self.marker = marker
-
-    def goal_response_callback(self, future):
-        goal_handle = future.result()
-        if not goal_handle.accepted:
-            self.get_logger().warn('Goal was rejected by Nav2.')
-            self.goal_sent = False
-            return
-
-        self._get_result_future = goal_handle.get_result_async()
-        self._get_result_future.add_done_callback(self.get_result_callback)
-        
-    def get_result_callback(self, future):
-        result = future.result().result
-            
-        if result.path: 
-            for pose_stamped in result.path.poses:
-                pose_stamped.pose.orientation.w = 0.0
-            # result.path.poses.reverse()
-            self.path_publisher.publish(result.path)
-        else:
-            self.get_logger().warn("No path returned in result")
-        self.goal_sent = False
 
     def publish_lookahead_circle(self, x, y):
         marker = Marker()
